Tests/LibWeb/Text/expected/css/getComputedStyle-pseudo-element.txt

#foo:
  color: rgb(128, 128, 128)
  background-color: rgb(255, 255, 0)
#foo::before:
  color: rgb(128, 128, 128)
  background-color: rgb(0, 255, 255)
#foo::after:
  content: "bye"
  color: rgb(0, 128, 0)
#foo::first-letter:
  color: rgb(255, 0, 0)
#foo::selection:
  background-color: rgb(0, 0, 255)
